+++ vitest.config.js
@@ -9,6 +9,8 @@ export default defineConfig({
       headless: true,
       instances: [
         { browser: 'chromium' },
+        { browser: 'firefox' },
+        { browser: 'webkit' },
       ],
     },
     include: ['test/specs/**/*.spec.js'],